photo-gallery-backend/requirements.txt
boto3
fastapi
Pillow
pydantic
uvicorn
minio
python-multipart